fix: simple refactor after review
