+++ docs/content/guide/components.rst
@@ -8,7 +8,7 @@ Components let you reuse template fragments (cards, headers, profiles) with prop
 Component folder and file routing
 ---------------------------------
 
-The directory name used for components (e.g. ``_components``) is set on **``DEFAULT_COMPONENT_BACKENDS``** as ``COMPONENTS_DIR``. The file router reads that same string when building URL scans so the folder is skipped and does **not** become a route segment. You may set ``COMPONENTS_DIR`` on a single ``DEFAULT_PAGE_BACKENDS`` entry only when that router must use a different folder name than the first component backend. Only this configured name is skipped (not every directory that starts with an underscore). See :doc:`file-router` for the routing side.
+The directory name used for components (e.g. ``_components``) is set on **``DEFAULT_COMPONENT_BACKENDS``** as ``COMPONENTS_DIR``. The file router always uses that value when building URL scans so the folder is skipped and does **not** become a route segment. Only this configured name is skipped (not every directory that starts with an underscore). See :doc:`file-router` for the routing side.
 
 Backends and settings
 ---------------------
@@ -16,8 +16,8 @@ Backends and settings
 Components are provided by backends, similar to the page router. In Django settings, use the top-level dict ``NEXT_FRAMEWORK`` with the key **``DEFAULT_COMPONENT_BACKENDS``**: a list of backend configs. Each item is a dict that is passed unchanged into the backend class constructor:
 
 - ``BACKEND`` (str) — dotted import path of the backend class (default for the built-in file backend is ``"next.components.FileComponentsBackend"``).
-- ``COMPONENTS_DIR`` (str, default ``"_components"`` in framework defaults) — folder name under each page tree where components are discovered and the name the file router skips during URL discovery. The router copies this value from the first component backend dict unless you override ``COMPONENTS_DIR`` on that page router entry.
-- ``DIRS`` (list) — extra filesystem directories registered as **global** component roots (visible from every template). Entries are split into real paths and segment names the same way as page ``DIRS``; only existing directory paths are used. App and root page trees do not need to be listed here because the same walk that builds URL patterns registers component folders there.
+- ``COMPONENTS_DIR`` (str, default ``"_components"`` in framework defaults) — folder name under each page tree where components are discovered and the name the file router skips during URL discovery. The file router reads this value from the first component backend entry.
+- ``DIRS`` (list) — extra filesystem directories registered as **global** component roots (visible from every template). Entries are split into real paths and segment names the same way as page ``DIRS``. Only existing directory paths are used. App and root page trees do not need to be listed here because the same walk that builds URL patterns registers component folders there.
 - You may list **several backends** in ``DEFAULT_COMPONENT_BACKENDS``. Earlier entries win when the same component name appears twice.
 
 ``component.py`` modules are always loaded with the framework’s built-in :class:`~next.components.ModuleLoader`.
@@ -69,7 +69,7 @@ Custom backends are plain classes referenced by dotted path in ``BACKEND``. Each
 Several roots in ``DEFAULT_PAGE_BACKENDS`` ``DIRS``
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-You may list more than one filesystem root under ``NEXT_FRAMEWORK`` ``DEFAULT_PAGE_BACKENDS`` ``DIRS``. Each entry that resolves to an existing directory is a separate page tree and is walked the same way as a single root setup. Set ``COMPONENTS_DIR`` once under ``DEFAULT_COMPONENT_BACKENDS`` unless a page backend entry needs its own override. The file router then skips that folder name when scanning routes.
+You may list more than one filesystem root under ``NEXT_FRAMEWORK`` ``DEFAULT_PAGE_BACKENDS`` ``DIRS``. Each entry that resolves to an existing directory is a separate page tree and is walked the same way as a single root setup. Set ``COMPONENTS_DIR`` on ``DEFAULT_COMPONENT_BACKENDS``. The file router uses that name when scanning every page tree.
 
 Components that live in ``<root>/<COMPONENTS_DIR>/`` at the top of a tree use the empty route scope for name resolution. That level is one shared namespace across all those roots. You must not reuse the same component name in the root component folder of two different roots. Running ``python manage.py check`` surfaces that situation as ``next.E034`` and lists the paths that collide.
