+++ docs/content/reference/system-checks.rst
@@ -58,7 +58,7 @@ Router list shape, backends, and ``FileRouterBackend`` fields.
      - ``DEFAULT_COMPONENT_BACKENDS`` must be a list
      - Error
    * - ``next.E024``–``E027``
-     - Missing or invalid ``PAGES_DIR`` / ``APP_DIRS`` / ``DIRS`` / ``COMPONENTS_DIR`` / ``OPTIONS`` for file router
+     - Missing or invalid ``PAGES_DIR`` / ``APP_DIRS`` / ``DIRS`` / ``OPTIONS`` for file router
      - Error
 
 **What to do:** Fix ``NEXT_FRAMEWORK`` in settings (see :doc:`../guide/file-router`).
@@ -191,7 +191,7 @@ See :doc:`../guide/context`.
 
 Flat top-level keys (see :mod:`next.conf`):
 
-* ``DEFAULT_PAGE_BACKENDS`` — list of file-router backend dicts (``BACKEND``, ``PAGES_DIR``, ``APP_DIRS``, ``DIRS``, ``OPTIONS``, …). Optional ``COMPONENTS_DIR`` overrides the folder name taken from ``DEFAULT_COMPONENT_BACKENDS``.
+* ``DEFAULT_PAGE_BACKENDS`` — list of file-router backend dicts (``BACKEND``, ``PAGES_DIR``, ``APP_DIRS``, ``DIRS``, ``OPTIONS``, …). The skip-folder name for routing comes from ``DEFAULT_COMPONENT_BACKENDS``, not from page router dicts.
 * ``URL_NAME_TEMPLATE`` — Python format string for URL names (default ``page_{name}``).
 * ``DEFAULT_COMPONENT_BACKENDS`` — list of component backend dicts.
