+++ examples/pages/config/settings.py
@@ -75,7 +75,6 @@ NEXT_FRAMEWORK = {
             "PAGES_DIR": "pages",
             "APP_DIRS": True,
             "DIRS": [],
-            "COMPONENTS_DIR": "_components",
             "OPTIONS": {
                 "context_processors": [],
             },
